1.2.0: box/furniture presence protocol, module lifecycle refactor, trade and UI fixes
Squashed 1.2.0 release changes since upstream/master.

Boxes and furniture share one authoritative possession protocol (pickup, throw,
placement, box-up/sell, join catch-up); sync modules run on a unified lifecycle
catalog/router; a client keeps worker-held boxes in the worker's hands and no
longer duplicates furniture or graded-card delivery boxes (a card box destroyed
locally now retires its host copy instead of re-spawning); trades carry
authoritative accept/refuse outcomes; worker-bonus, shelf-pull and box-visual
updates are immediate; and the co-op window gains a SETTINGS tab for diagnostics.
See CHANGELOG.md for the player-facing summary.

+++ src/CardShopCoop/Net/Messages/SystemMessages.cs
@@ -49,7 +49,8 @@ namespace CardShopCoop.Net.Messages
         public int SaveLength;
         public int HostSlot;
         public int BundleLength;
-        public byte SelfId;
+        // int, not byte: the host's connection ids are unbounded; a byte wrapped at 255.
+        public int SelfId;
         public byte[] HostEnumBlob = new byte[0];
         public byte[] HostCardsBlob = new byte[0];
         public ulong SteamId;
